moved the get_commitment to query_server and used ledger_cloned there
