+++ src/components/abciapp/src/api/query_server/query_api/server.rs
@@ -2,6 +2,8 @@
 //! data sources for the query api
 //!
 
+use globutils::HashOf;
+use ledger::data_model::StateCommitmentData;
 use {
     lazy_static::lazy_static,
     ledger::{
@@ -300,6 +302,23 @@ impl QueryServer {
             }
         }
         memos
+    }
+
+    #[inline(always)]
+    #[allow(missing_docs)]
+    pub fn get_state_commitment_from_api_cache(
+        &self,
+    ) -> (HashOf<Option<StateCommitmentData>>, u64) {
+        let block_count = self.ledger_cloned.get_block_commit_count();
+        let commitment = self
+            .ledger_cloned
+            .api_cache
+            .as_ref()
+            .unwrap()
+            .state_commitment_version
+            .clone()
+            .unwrap_or_else(|| HashOf::new(&None));
+        (commitment, block_count)
     }
 
     /// Returns the abar commitment by given index, if it exists.
